requirements.txt
numpy
torch
torchvision
pillow
gdspy
matplotlib
opencv-python
scipy
pandas
tqdm